refactor: deepen viewer delivery module
