+++ src/server/mod.rs
@@ -23,13 +23,10 @@ use axum::{
     Router,
 };
 use crate::protocol;
-use bytes::Bytes;
-use futures_util::{SinkExt, StreamExt};
 use binaries::BinaryDownloadQuery;
+use bytes::Bytes;
 use rooms::{RoomId, Rooms};
-use std::collections::HashMap;
 use std::future::IntoFuture;
-use std::sync::{Arc, OnceLock};
 use std::time::Duration;
 use tower_http::trace::TraceLayer;
 use tracing::{debug, info, warn};
@@ -53,19 +50,12 @@ impl Default for CleanupConfig {
 }
 
 /// Shared application state
-#[derive(Default, Clone)]
+#[derive(Clone)]
 struct AppState {
     /// All live rooms (state, passwords, history, eviction)
     rooms: Rooms,
-    /// Queues feeding the fan-out tasks, one per shard - set once at
-    /// startup. A room always hashes to the same shard, so
-    /// per-room ordering holds; spreading rooms across shards keeps one
-    /// busy room from delaying every other room's viewers
-    fanout: Arc<OnceLock<Vec<tokio::sync::mpsc::UnboundedSender<FanoutItem>>>>,
-    /// Rooms whose user count changed and needs re-broadcasting
-    usercount: Arc<OnceLock<tokio::sync::mpsc::UnboundedSender<String>>>,
-    /// Raw-WebSocket viewers, by room
-    viewers: viewers::Viewers,
+    /// Raw-WebSocket viewer delivery, including fan-out and membership
+    viewer_delivery: viewers::ViewerDelivery,
     /// Cleanup configuration for abandoned rooms
     cleanup_config: CleanupConfig,
     /// Optional usage analytics; a no-op unless the operator opted in
@@ -149,35 +139,21 @@ pub async fn serve_on(
         cleanup_interval_secs, room_ttl_secs
     );
 
-    // Shared state with cleanup configuration
+    // Shared state with cleanup configuration. Viewer delivery is mandatory:
+    // constructing it here removes the invalid "server is running but its
+    // fan-out queues were never installed" state.
+    let rooms = Rooms::default();
+    let analytics = analytics::Analytics::new(analytics_config);
+    let viewer_delivery = viewers::ViewerDelivery::start(rooms.clone(), analytics.clone());
     let app_state = AppState {
+        rooms,
+        viewer_delivery,
         cleanup_config: CleanupConfig {
             interval: Duration::from_secs(cleanup_interval_secs),
             inactive_ttl: Duration::from_secs(room_ttl_secs),
         },
-        analytics: analytics::Analytics::new(analytics_config),
-        ..Default::default()
+        analytics,
     };
-
-    // The viewer fan-out tasks: ingest stores and acks, these emit.
-    // The tasks get only the viewer registry, not AppState: AppState
-    // holds their senders, and a task holding its own sender would keep
-    // its channel (and itself) alive forever. One task per shard, sized
-    // to the machine, so concurrent rooms emit in parallel
-    let shards = std::thread::available_parallelism().map_or(4, std::num::NonZeroUsize::get);
-    let mut fanout_txs = Vec::with_capacity(shards);
-    for _ in 0..shards {
-        let (tx, rx) = tokio::sync::mpsc::unbounded_channel();
-        fanout_txs.push(tx);
-        tokio::spawn(fanout_loop(rx, app_state.viewers.clone()));
-    }
-    let _ = app_state.fanout.set(fanout_txs);
-
-    // The user-count broadcast task: joins and disconnects queue the
-    // room here instead of broadcasting inline
-    let (usercount_tx, usercount_rx) = tokio::sync::mpsc::unbounded_channel();
-    let _ = app_state.usercount.set(usercount_tx);
-    tokio::spawn(usercount_loop(usercount_rx, app_state.viewers.clone()));
 
     // Spawn background cleanup task for abandoned rooms
     spawn_cleanup_task(app_state.clone());
@@ -260,160 +236,9 @@ fn ingest(
     message: Option<&Bytes>,
 ) -> Result<(), rooms::Unauthorized> {
     let _ = state.rooms.append(room_id, secret, size, message)?;
-
-    if let Some(txs) = state.fanout.get() {
-        // The channels are unbounded; depth is bounded in practice by
-        // the aggregate ingest rate across broadcasters, and queued
-        // payloads are refcounted slices of already-stored history. A
-        // room always lands on the same shard (hash below), which is
-        // what preserves its ordering; rooms sharing a shard can still
-        // delay each other, but the blast radius is 1/shards
-        let shard = room_shard(room_id.as_str(), txs.len());
-        let _ = txs[shard].send(FanoutItem {
-            room: room_id.as_str().to_string(),
-            size: size.cloned(),
-            payload: message.cloned(),
-        });
-    }
+    state.viewer_delivery.publish_ingest(room_id, size, message);
 
     Ok(())
-}
-
-/// Stable room -> fan-out shard assignment.
-fn room_shard(room: &str, shards: usize) -> usize {
-    use std::hash::{Hash, Hasher};
-    let mut hasher = std::collections::hash_map::DefaultHasher::new();
-    room.hash(&mut hasher);
-    usize::try_from(hasher.finish() % shards.max(1) as u64).unwrap_or(0)
-}
-
-/// One unit of viewer fan-out work: a `size` control event and/or a
-/// binary terminal payload for a room, in broadcast order.
-struct FanoutItem {
-    room: String,
-    size: Option<serde_json::Value>,
-    payload: Option<Bytes>,
-}
-
-/// Max bytes coalesced into a single viewer frame. Matches the client's
-/// own send batching (`MAX_BATCH` in `cli/script.rs`), so viewers never
-/// see a frame shape the client couldn't already have produced.
-const FANOUT_MAX_BATCH: usize = 64 * 1024;
-
-/// The viewer fan-out task: drains the queue and emits to viewers.
-///
-/// Whatever queued up while the previous emits ran is coalesced - each
-/// room's payloads are concatenated (up to [`FANOUT_MAX_BATCH`] per
-/// emit), exactly like the client's sender thread coalesces PTY output.
-/// Under burst load this collapses thousands of tiny per-socket sends
-/// into a few large ones, which is what keeps slow viewers' buffers
-/// from overflowing into silent content loss. Terminal output is a raw
-/// byte stream of whole frames, so concatenation is invisible to
-/// viewers (room history already concatenates the same way).
-///
-/// A single task consumes the queue, so per-room ordering is exactly
-/// the ingest order (cross-room ordering carries no meaning). The
-/// store-then-queue gap means a viewer joining mid-burst may see a
-/// queued frame around its history replay - duplicated, or even before
-/// the replay containing it. That race pre-exists this task (emits were
-/// always concurrent with the join handler); the queue widens it from
-/// microseconds to the drain latency. Same class as the duplicate
-/// render already accepted around client reconnect replay: delivery is
-/// at-least-once end to end.
-async fn fanout_loop(
-    mut rx: tokio::sync::mpsc::UnboundedReceiver<FanoutItem>,
-    viewers: viewers::Viewers,
-) {
-    /// Payloads accumulated for one room, flushed as one emit.
-    #[derive(Default)]
-    struct Pending {
-        chunks: Vec<Bytes>,
-        len: usize,
-    }
-
-    fn flush(viewers: &viewers::Viewers, room: &str, p: Pending) {
-        if p.chunks.is_empty() {
-            return;
-        }
-        let payload = if p.chunks.len() == 1 {
-            p.chunks.into_iter().next().unwrap_or_default()
-        } else {
-            let mut buf = bytes::BytesMut::with_capacity(p.len);
-            for chunk in p.chunks {
-                buf.extend_from_slice(&chunk);
-            }
-            buf.freeze()
-        };
-        viewers.send_bytes(room, &payload);
-    }
-
-    while let Some(first) = rx.recv().await {
-        let mut batch = vec![first];
-        while let Ok(item) = rx.try_recv() {
-            batch.push(item);
-        }
-        let mut pending: HashMap<String, Pending> = HashMap::new();
-        for item in batch {
-            if let Some(size) = item.size {
-                // A size event is an ordering barrier within its room:
-                // anything queued before it must reach viewers first
-                if let Some(p) = pending.remove(&item.room) {
-                    flush(&viewers, &item.room, p);
-                }
-                let control = serde_json::json!({ "size": size }).to_string();
-                viewers.send_control(&item.room, &control);
-            }
-            if let Some(payload) = item.payload {
-                match pending.entry(item.room) {
-                    std::collections::hash_map::Entry::Occupied(mut e) => {
-                        if e.get().len + payload.len() > FANOUT_MAX_BATCH {
-                            let full = std::mem::take(e.get_mut());
-                            flush(&viewers, e.key(), full);
-                        }
-                        let p = e.get_mut();
-                        p.len += payload.len();
-                        p.chunks.push(payload);
-                    }
-                    std::collections::hash_map::Entry::Vacant(v) => {
-                        v.insert(Pending {
-                            len: payload.len(),
-                            chunks: vec![payload],
-                        });
-                    }
-                }
-            }
-        }
-        for (room, p) in pending {
-            flush(&viewers, &room, p);
-        }
-    }
-}
-
-/// The user-count broadcast task: re-announces a room's viewer count
-/// to the room whenever its membership changed.
-///
-/// Joins and disconnects queue the room name here instead of
-/// broadcasting inline: an audience of N joining produces N broadcasts
-/// to up to N members - O(N^2) emits in a connect storm. Draining and
-/// deduplicating turns that into at most one broadcast per room per
-/// pass, each carrying the count current at emit time, so every viewer
-/// still converges on the exact final number (intermediate values may
-/// be skipped, exactly as if the joins had raced the same broadcast).
-async fn usercount_loop(
-    mut rx: tokio::sync::mpsc::UnboundedReceiver<String>,
-    viewers: viewers::Viewers,
-) {
-    let mut rooms = std::collections::HashSet::new();
-    while let Some(first) = rx.recv().await {
-        rooms.insert(first);
-        while let Ok(room) = rx.try_recv() {
-            rooms.insert(room);
-        }
-        for room in rooms.drain() {
-            let count = viewers.count(&room);
-            viewers.send_control(&room, &format!("{{\"usersCount\":{count}}}"));
-        }
-    }
 }
 
 /// GET /ws/v/r/:room - raw-WebSocket viewer endpoint.
@@ -431,213 +256,8 @@ async fn ws_view_handler(
     ws: WebSocketUpgrade,
 ) -> Response {
     let room_id = RoomId::parse(&room_path);
-    ws.on_upgrade(move |socket| ws_view_loop(socket, state, room_id))
-}
-
-/// How long a viewer may go without sending anything (pong frames
-/// answer our pings automatically in every browser) before the
-/// connection is presumed dead. Pings go out every 25s, so a healthy
-/// peer is never close to this. The same bound caps every write: a
-/// peer that stops reading (frozen tab, zero TCP window) would
-/// otherwise block `sink.send` forever, pinning the task and its
-/// backlog - the select loop can't reach its idle check while a send
-/// is in flight.
-const VIEWER_IDLE_TIMEOUT: Duration = Duration::from_secs(75);
-
-/// Send with [`VIEWER_IDLE_TIMEOUT`] as the stall bound.
-async fn timed_send<S>(sink: &mut S, msg: WsMessage) -> Result<(), axum::Error>
-where
-    S: futures_util::Sink<WsMessage, Error = axum::Error> + Unpin,
-{
-    match tokio::time::timeout(VIEWER_IDLE_TIMEOUT, sink.send(msg)).await {
-        Ok(result) => result,
-        Err(elapsed) => Err(axum::Error::new(elapsed)),
-    }
-}
-/// Viewer ping cadence; keeps NATs open and detects dead peers.
-const VIEWER_PING_INTERVAL: Duration = Duration::from_secs(25);
-
-/// One viewer connection: replay the catch-up snapshot, then relay the
-/// registry queue until the peer leaves or stalls out.
-async fn ws_view_loop(mut socket: WebSocket, state: AppState, room_id: RoomId) {
-    let room = room_id.as_str().to_string();
-    // Register BEFORE the snapshot: a frame broadcast in between shows
-    // up in the queue (and possibly also in the snapshot - the
-    // accepted at-least-once duplicate), but can never be missed
-    let (viewer_id, mut rx) = state.viewers.join(&room);
-    info!("WS viewer {viewer_id} joined room {room_id:?}");
-
-    let snapshot = state.rooms.snapshot(&room_id);
-    let broadcasting = snapshot.as_ref().is_some_and(|s| s.broadcasting);
-    let room_exists = snapshot.is_some();
-    let catch_up = async {
-        if let Some(snap) = snapshot {
-            if let Some(size) = snap.size {
-                let frame = serde_json::json!({ "size": size }).to_string();
-                timed_send(&mut socket, WsMessage::Text(frame.into())).await?;
-            }
-            if let Some(history) = snap.history {
-                timed_send(&mut socket, WsMessage::Binary(history)).await?;
-            }
-        }
-        let frame = format!("{{\"broadcasting\":{broadcasting}}}");
-        timed_send(&mut socket, WsMessage::Text(frame.into())).await?;
-        let count = total_viewers(&state, &room);
-        timed_send(
-            &mut socket,
-            WsMessage::Text(format!("{{\"usersCount\":{count}}}").into()),
-        )
-        .await
-    };
-    if catch_up.await.is_err() {
-        state.viewers.leave(&room, viewer_id);
-        // The room may have seen this viewer in a count broadcast
-        // during its brief membership; converge the others
-        if let Some(tx) = state.usercount.get() {
-            let _ = tx.send(room);
-        }
-        return;
-    }
-    // The rest of the room learns the new count via the coalescing task
-    if let Some(tx) = state.usercount.get() {
-        let _ = tx.send(room.clone());
-    }
-    // Joins to nonexistent rooms (dead links) are not an audience
-    if room_exists {
-        state
-            .analytics
-            .viewer_joined(room_id.as_str(), total_viewers(&state, &room), broadcasting);
-    }
-
-    // Split the socket: the writer must never poll the read half. A
-    // read poll per delivered message costs a wasted read syscall and
-    // tungstenite read/write interplay - measured as ~4x the delivery
-    // latency at thousands of viewers. The reader task only proves the
-    // peer is alive (pongs and anything else it sends) and reports
-    // when the connection dies.
-    let (mut sink, mut stream) = socket.split();
-    let started = tokio::time::Instant::now();
-    let last_seen_ms = Arc::new(std::sync::atomic::AtomicU64::new(0));
-    let seen = last_seen_ms.clone();
-    let mut reader = tokio::spawn(async move {
-        while let Some(Ok(_)) = stream.next().await {
-            let elapsed_ms = started.elapsed().as_millis().min(u128::from(u64::MAX));
-            #[allow(clippy::cast_possible_truncation)] // bounded by the min above
-            seen.store(elapsed_ms as u64, std::sync::atomic::Ordering::Relaxed);
-        }
-    });
-    let mut ping = tokio::time::interval(VIEWER_PING_INTERVAL);
-    ping.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Delay);
-    loop {
-        tokio::select! {
-            queued = rx.recv() => {
-                // Dropped by the registry for stalling: close, so the
-                // page reconnects and resyncs from history
-                let Some(first) = queued else { break };
-                if relay(&mut sink, &mut rx, first).await.is_err() {
-                    break;
-                }
-            }
-            // The peer hung up (or errored): stop delivering
-            _ = &mut reader => break,
-            _ = ping.tick() => {
-                let last_seen = Duration::from_millis(
-                    last_seen_ms.load(std::sync::atomic::Ordering::Relaxed),
-                );
-                if started.elapsed().saturating_sub(last_seen) > VIEWER_IDLE_TIMEOUT {
-                    info!("WS viewer {viewer_id} in {room_id:?} timed out");
-                    break;
-                }
-                if timed_send(&mut sink, WsMessage::Ping(Vec::new().into()))
-                    .await
-                    .is_err()
-                {
-                    break;
-                }
-            }
-        }
-    }
-    reader.abort();
-    // Best effort: a Close frame lets the page distinguish a server
-    // close from a network drop (it reconnects either way)
-    let _ = tokio::time::timeout(
-        Duration::from_secs(1),
-        sink.send(WsMessage::Close(None)),
-    )
-    .await;
-    state.viewers.leave(&room, viewer_id);
-    info!("WS viewer {viewer_id} left room {room_id:?}");
-    if let Some(tx) = state.usercount.get() {
-        let _ = tx.send(room);
-    }
-}
-
-/// Relay one wake-up's worth of queued messages to the viewer.
-///
-/// Everything already queued is drained and consecutive binary
-/// payloads are merged into one WebSocket frame (terminal output is a
-/// raw byte stream, so the merge is invisible - the client batches the
-/// same way before sending). Every message otherwise costs its own
-/// write+flush, and at thousands of viewers x dozens of frames/s those
-/// per-frame flushes dominate the server. Control events keep their
-/// own text frames, and ordering is preserved throughout.
-async fn relay(
-    sink: &mut futures_util::stream::SplitSink<WebSocket, WsMessage>,
-    rx: &mut tokio::sync::mpsc::Receiver<viewers::ViewerMsg>,
-    first: viewers::ViewerMsg,
-) -> Result<(), axum::Error> {
-    /// Bound on a merged frame; beyond it the backlog flushes in parts.
-    const MAX_MERGED: usize = 1024 * 1024;
-    /// One frame from the accumulated chunks - without copying when a
-    /// single chunk stands alone (the common case at normal pace).
-    fn take_merged(chunks: &mut Vec<Bytes>, len: &mut usize) -> Bytes {
-        *len = 0;
-        if chunks.len() == 1 {
-            return chunks.pop().unwrap_or_default();
-        }
-        let mut buf = bytes::BytesMut::with_capacity(chunks.iter().map(Bytes::len).sum());
-        for chunk in chunks.drain(..) {
-            buf.extend_from_slice(&chunk);
-        }
-        buf.freeze()
-    }
-    let mut chunks: Vec<Bytes> = Vec::new();
-    let mut pending_len = 0;
-    let mut item = Some(first);
-    loop {
-        match item {
-            Some(viewers::ViewerMsg::Bytes(payload)) => {
-                if pending_len + payload.len() > MAX_MERGED && !chunks.is_empty() {
-                    let merged = take_merged(&mut chunks, &mut pending_len);
-                    timed_send(sink, WsMessage::Binary(merged)).await?;
-                }
-                pending_len += payload.len();
-                chunks.push(payload);
-            }
-            Some(viewers::ViewerMsg::Control(json)) => {
-                if !chunks.is_empty() {
-                    let merged = take_merged(&mut chunks, &mut pending_len);
-                    timed_send(sink, WsMessage::Binary(merged)).await?;
-                }
-                timed_send(sink, WsMessage::Text(json.into())).await?;
-            }
-            None => break,
-        }
-        item = match rx.try_recv() {
-            Ok(next) => Some(next),
-            Err(_) => break,
-        };
-    }
-    if !chunks.is_empty() {
-        let merged = take_merged(&mut chunks, &mut pending_len);
-        timed_send(sink, WsMessage::Binary(merged)).await?;
-    }
-    Ok(())
-}
-
-/// How many viewers are watching the room right now.
-fn total_viewers(state: &AppState, room: &str) -> usize {
-    state.viewers.count(room)
+    let delivery = state.viewer_delivery;
+    ws.on_upgrade(move |socket| async move { delivery.join(room_id, socket).await })
 }
 
 /// GET /ws/r/:room - WebSocket ingest for broadcasting clients.
@@ -798,8 +418,8 @@ async fn ws_ingest_loop(
                         break;
                     }
                     state
-                        .viewers
-                        .send_control(room_id.as_str(), "{\"reset\":true}");
+                        .viewer_delivery
+                        .publish_control(&room_id, viewers::ViewerControl::Reset);
                     continue;
                 }
                 let size = body
@@ -871,8 +491,8 @@ async fn recv_with_timeout(
 /// Tell every viewer in the room whether a broadcaster is attached
 fn emit_broadcasting(state: &AppState, room_id: &RoomId, live: bool) {
     state
-        .viewers
-        .send_control(room_id.as_str(), &format!("{{\"broadcasting\":{live}}}"));
+        .viewer_delivery
+        .publish_control(room_id, viewers::ViewerControl::Broadcasting(live));
 }
 
 /// GET /r/:room - the viewer page, or - with a `.bin` suffix - the
